Move code to ProcessPicker component
